+++ cmd/query-tee/main_test.go
@@ -8,12 +8,12 @@ import (
 )
 
 func TestCortexReadRoutes(t *testing.T) {
-	routes := cortexReadRoutes("")
+	routes := cortexReadRoutes(Config{PathPrefix: ""})
 	for _, r := range routes {
 		assert.True(t, strings.HasPrefix(r.Path, "/api/v1/"))
 	}
 
-	routes = cortexReadRoutes("/some/random/prefix///")
+	routes = cortexReadRoutes(Config{PathPrefix: "/some/random/prefix///"})
 	for _, r := range routes {
 		assert.True(t, strings.HasPrefix(r.Path, "/some/random/prefix/api/v1/"))
 	}